requirements.txt
langchain
python-dotenv
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain-ollama
torch
transformers
langchain_groq
langchain_core
langserve
fastapi
uvicorn
sse_starlette
ddgs
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
yt-dlp
numexpr
sympy
huggingface_hub
cassio
datasets
tiktoken
openai
pyPDF2
gradio
langchain-nvidia-ai-endpoints==0.3.19
crewai
crewai_tools
qdrant-client
crewai[rags_chromadb]
crewai[litellm]
pinecone
pinecone-text
pinecone-notebooks
neo4j==5.14
langgraph
langsmith